@@ -8,7 +8,7 @@ module Amalgamate
     attr_accessor :master, :slave
     def initialize(*args)
       @master, @slave = args
-      raise ClassMismatchError unless @master.class.eql?(@slave.class) 
+      raise ClassMismatchError unless @master.class.eql?(@slave.class)
     end
 
     def unify(*args)
@@ -24,11 +24,11 @@ module Amalgamate
         memo[attribute] = values[priority].nil? ? values[secondary] : values[priority]
         memo
       end
-      master.assign_attributes(merge_values, without_protection: true)
+      master.assign_attributes(merge_values)
       master.save if options[:save] != false && master.changed?
       self.reassign_associations(master, slave, priority: priority) unless options[:reassign_associations] == false
       slave.destroy unless options[:destroy] == false
-      master 
+      master
     end
 
     def diff(*args)
@@ -36,7 +36,7 @@ module Amalgamate
       master, slave = args if args.any?
       master ||= @master
       slave ||= @slave
-      options[:ignore] ||= IGNORED_ATTRIBUTES 
+      options[:ignore] ||= IGNORED_ATTRIBUTES
       master.attributes.keys.inject({}) do |memo, attribute|
         attribute = attribute.to_sym
         unless options[:ignore].include?(attribute)
