@@ -1,3 +1,3 @@
 module Amalgamate
-  VERSION = "0.1.0"
+  VERSION = "0.1.1"
 end
